requirements.txt
django==2.0.7
Pillow
exifread
django-imagekit